Changed VehicleStatusTests to EmergencyStopTest and removed tests that link to nonexistent posts and gets
